+++ features/step_definitions/cookbook_steps.rb
@@ -1,29 +1,3 @@
-Given /^a cookbook with a single recipe that accesses node attributes via strings$/ do
-  write_recipe %q{node['foo'] = 'bar'}
-end
-
-Given /^a cookbook with a single recipe that accesses multiple node attributes via strings$/ do
-  write_recipe %q{node['foo'] = 'bar'
-node['testing'] = 'bar'
-  }
-end
-
-Given /^a cookbook with a single recipe that assigns node attributes accessed via strings to a local variable$/ do
-  write_recipe %q{baz = node['foo']}
-end
-
-Given /^a cookbook with a single recipe that accesses nested node attributes via strings$/ do
-  write_recipe %q{node['foo']['foo2'] = 'bar'}
-end
-
-Given /^a cookbook with a single recipe that accesses node attributes via symbols/ do
-  write_recipe %q{node[:foo] = 'bar'}
-end
-
-Given /^a cookbook that declares ([a-z]+) attributes via strings$/ do |attribute_type|
-  write_attributes %Q{#{attribute_type}["apache"]["dir"] = "/etc/apache2"}
-end
-
 When /^I check the cookbook$/ do
   run_lint
 end