+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0 (1st December, 2011)
+
+Bugfixes:
+
+  - Removed 'FC001: Use symbols in preference to strings to access node attributes' until a policy mechanism is
+  introduced ([related issue](https://github.com/acrmp/foodcritic/issues/1)). Thanks @jtimberman
+
 ## 0.1.0 (30th November, 2011)
 
 Initial version.